requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyodbc==5.2.0
pyspnego==0.11.2
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
smbprotocol==1.15.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typer==0.14.0
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
